Add ESLint configuration file for docs
The reasoning for this change is to get rid of ESLint errors shown in the IDEs
that use ESLint plugins (e.g. VSCode) because the ESLint default rules for importing
extraneus modules are not observed in docs app - it uses the modules that are not
listed in main content dependencies section.

Thanks!