fix(CarouselIndicators): li key values generate from provided item values
Generation of key items for `li` components in `CarouselIndicators` is a
concatenation of specific, optional values (`src`, `caption`, and
`altText`) in each item of the `items` object array. Lacking these
key/value pairs will cause each `li` element to have a key along the
lines of `undefinedundefinedundefined`, resulting two or more of the
`li`'s to be identical and causing an error to log in the console.
While either including any one of the specified keys with a unique
value, or adding a unique `key` field in each of the items objects will
remedy this issue, that solution is not well documented. The proposed
solution is a transparent fix: preventing the aforementioned error while
requiring no change of code for the end-user. The new `key` prop value
for each `li` will be a concatenation of all the values of its
respective item object.

Closes #1310