Restore legacy Popover behavior via `trigger="legacy-auto-hide" isInteractive delay={0}`
The new trigger hides the popover when clicking anywhere except the target element
and, depending on the new `isInteractive` prop, the popover itself.
It can be combined with the `click` trigger, in which case the popover is shown
when clicking the target element and hidden when clicking somewhere else.

Tackles issue #1349.