Convert Form and Input to stateless components.
These components only needed to be a class component because it had the getRef
and focus/submit instance methods. getRef seems to be an undocumented vestige from
an earlier implementation. It is also the only method by which this.ref
could get set on the instance, which is what focus/submit would use. Removing getRef
requires removing focus/submit, but this is okay because converting it to a
functional component means that forwardRef will be able to supply the native
DOM element that has the desired function.