feat(Dropdown, DropdownMenu, DropdownToggle): allow Dropdown to have listbox or menu role
Dropdown Menu should allow for listbox or menu roles.  Before this commit, the Dropdown could only have the menu role.  After this commit, a user can add a menuRole property to the Dropdown component which will set a corresponding aria-haspopop property on the DropdownToggle, role property on the DropdownMenu, and role property on the Dropdown Items